Improvements to OculusViewConfig to improve interoperability with 3rd Party OSG applications
Refactored OculusViewConfig so that is uses
osg::View::Slave::SlaveUpdateCallback rather than osg::Camera cull
callback.  Using the slave callback avoids the need for assigning the
graphics context to the View's master Camera and with it the need for
setting a NodeMask to avoid rendering the master Camera's subgraph. This
change means that client applications won't need to alter there NodeMask
usage to use OculusViewConfig.